Implicit creation of networks if the net mode is driver:nameOrId
Relax the treatment of the `--net` flag so that it can specify a
driver and a network name; e.g., `--net=driver:nameOrId`. The network
will be created with the specified driver, if it doesn't already
exist.

This is a simple way of being able to create a network in libnetwork
via docker.

Conflicts:
	daemon/container.go